修改了一些函数原型，使用int代替uint64_t作为key的类型，同时完成二叉搜索辅助函数

+++ Btree/btree.h
@@ -30,8 +30,8 @@ struct tree{
 };
 #define node_pointer TOID(struct tree_node)
 //Output
-node_pointer find_leaf(node_pointer root,uint64_t key);
-PMEMoid find(node_pointer root,uint64_t key);
+node_pointer find_leaf(node_pointer root,int key);
+PMEMoid find(node_pointer root,int key);
 //Insertion
 PMEMoid make_node();
 PMEMoid make_leaf();
